[backport] Use a stricter pin for ffi-yajl (#118)
Backport of #117.

ffi-yajl 2.4.0 was released a few hours ago: https://rubygems.org/gems/ffi-yajl/versions/2.4.0-universal-java. This version changes the libyajl2 dependency to a more lenient one (~> 1.2 became >= 1.2), which is currently breaking Windows builds.

This PR reverts the dependency to its previous version for now (to fix builds on master & for the current release). Properly updating ffi-yajl will be done in a future PR.